More bit hack improvements.
No longer do a modulo in the macros when not needed,
but require calling code to do that.

BITMASK wasn't used. I redefined it and now use it
to create a mask with just one bit set for one column,
and use it everywhere where before ONE was shifted.

Improved documentation of LEFT_BITMASK and RIGHT_BITMASK,
and use them in packedmatrix.h instead of explicitely
shifting ONE there.