More micro optimizations and a bug fix.
Removed LEFTMOST_BITS and RIGHTMOST_BITS completely:
the meaning 'left' and 'right' is a bit fuzzy at
the moment ;)

Whereever we do an AND with ~0xF, use ~0xFUL instead,
although sign extention is saving us, it's a bit
scary to interpret 0xF as (signed) int, invert it
to get 0xFFFFFFF0 and then because it is signed
get 0xFFFFFFFFFFFFFFF0 before it's assigned to an
unsigned long (long).

Some micro optimization mostly involving the fact
that we can swap bits by doing an XOR with the XOR
of the bits that have to be swapped.

Fixed a bug in mzd_row_add_offset where non-zero
excess bits in the source would end up in the destination.
Apparently the destination is never a window into
a larger matrix, so this wasn't discovered by the
testsuite, but it might matter in cases where someone
relies on those excess bits to be zero.
While I was add it, I more or less rewrote the
function to be faster for smaller matrices without
being a clock cylce slower for larger ones ;)

Fixed several typo fixes in documentation of xor.h.