Introduction of MIDDLE_BITMASK.
MIDDLE_BITMASK is just 'LEFT_BITMASK & RIGHT_BITMASK' when operating
on the same word (thus, small matrices) but a little bit faster.

Also, set RADIX to just '64'. While 'sizeof(word) << 3' results in the
same it, 1) has type 'size_t' *) and 2) only works when CHAR_BIT == 8,
which is not the case on all machines :p.

Since we use uint64_t as type for word, it is more logical to use 64
explicitely than to pretend that word could be changed to a different
type (which it cannot)

(*) we shouldn't use size_t for bit sizes... but I'll get back to
that later).