Bug fix for _mzd_transpose_direct
I was incorrectly assuming that the rowstride is constant;
it is not when we enter a new block. This fix no doubt will
make transposing slower again, but we can't fix that until
after 'rowstride' is introduced.

Also, use ONE instead of 1 where appropriate, and use size_t
instead of word for bit distances.