Introduction of M4RI_WORDWRAP and the C++ class word.
This patch turns word into a C++ class that allows
extensive testing and checks, if the library is
configured with a C++ compiler as compiler.

For example:
CFLAGS="-O2" CC="g++" ./configure --enable-maintainer-mode --enable-debug

It is not possible to pass --enable-sse2, as the class contains
more members than just the underlaying uint64_t and therefore
no 128-bit operations are possible anymore.