Reversed the bit order of the internal representation of class word.
On top of that, remove operators <, >, <= and >=.
Comparing integer values for less than or greater than has to happen
by first converting to the underlaying uint64_t (using CONVERT_TO_UINT64_T).

This patch already exports the reversal of the bits through
CONVERT_TO_UINT64_T, so that functions that used operator< needed
to be reimplemented for the reversed case (larger_log2 --> lesser_LSB).
However, the reversal on conversion from uint64_t to word is still hidden,
breaking the demand that CONVERT_TO_UINT64_T(CONVERT_TO_WORD(val)) == val.
This will be fixed in the next commit.