Remove last traces of reverse.
This patch reverses the k bits of Gray codes of k bits long:
code::ord. It also moves those k bits to the other side
of the word when read with mzd_read_bits and also adjusts
mzd_xor_bits to work on the other side.

This means that the index of all L*, M* and E* array's
that take a Gray code as index are reversed (for their
respective lengths).

It was discovered that a test to break a loop in mzd_find_pivot
was unnecessary, as it was testing a bit that is always zero.
This test was removed (if we left it in we would have needed
to check bit index -1 of a word).

The library now compiles and works again when compiled with
a C compiler!