Type and whitespace clean up.
Introduction of rci_t and wi_t.

rci_t is the type of row or column index, and differences thereof.
However, when it is known the certain row/col difference is
significantly small, the type int is used. For example:

int skip = col % RADIX;

wi_t is the type of word index, and differences thereof.
This the type of differences between two word*.
Most notably, rci_t / RADIX = wi_t.

Note the types in mzd_t::rows[rci_t][wi_t].

This patch also moves a lot (all?) variable declaration from the
beginning of function to where they are actually (first) used.
This is supported by C99 and makes the code a lot more clear.

This patch also includes a manual re-indent, most notably putting
const on the right-hand side of types, using pre-increment when
possible and adding spaces around operators.

Finally, this patch contain temporary code to check correct usage
of rci_t and wi_t relative to int (size_t) and to eachother. This is
achieved by, next to word, also wrapping multiples of RADIX in a C++
class, as Radix_t, using wordPtr and wordPtrPtr as wrappers for
respectively word* and word** and by wrapping rci_t and wi_t in
C++ classes. This commit contains this code for archival purposes,
as it will be removed again in the next patch (this code only
compiles with a C++ compiler at the moment).