Bug fix in m4ri_random_word.
m4ri_random_word accidently only returned 31
bits instead of 64.

Improved output format of mzd_print to print
a '|' between words instead of a space, also
when offset != 0 (before a colon was printed
in that case). And to not print a trailing ':'
after the last bit.

Fixed mzd_equal (now returning int instead
of BIT), to also work for windowed matrices.

Fixed mzd_randomize to generate the same
matrix (after a call to srandom() with the same
seed) independent of the offset of a matrix,
and without overwriting the excess bits. In
other words, this functions now also works
for windowed matrices.