Prefix all exported variables, functions and macros.
Macros are now prefixed with __M4RI_, which the exception of MAX and MIN.
Functions are prefixed with m4ri_, mzd_ or _mzd.
Constants defined in headers are lower case and prefixed with m4ri_.
This doesn't make the code more readable unfortunately, but it's just
not professional to export things in a library that can easily collide with others.