More constness fixes.
This should cause all pointers passed to functions
to be a pointer-to-const when the content is not changed.

I needed to introduce mzd_init_window_const, which creates
a window into a const mzd_t, returning mzd_t const* as well.
I decided to demand an explicit cast when freeing such
a window (we can fix that later once I added flags to mzd_t,
and add runtime checking when freeing a const window,
removing the need for explicit casts).

Also fixed a few -Wall compiler warnings that sneaked
into the testsuite.