Add new elements to mzd_t and keep them consistent.
After this patch it is  guaranteed that excess bits and padding
are zero (I think they already were, but now I checked it, see
next commit).

rowstride is the offset between two rows within a block.
blockrows_log and blockrows_mask are respectively the log2
of the number of rows in blocks before the last block,
and the number of rows minus one. Note that number of rows
is exactly a power of 2: 1 << blockrows_log, so that given
some row, the first word in that row is given by:
blocks[row >> blockrows_log] + (row & blockrows_mask) * rowstride.

high_bitmask and low_bitmask are precomputed bitmasks, masking
valid bits in the case of excess bits and/or offset bits respectively.
If width == 1, then both are equal and mask all valid bits.

Matrices with a width less than mzd_paddingwidth (currently 3),
are no longer 128-bit aligned. This makes small matrices (width = 1),
twice as compact while they already weren't processed with SSE2
instruction anyway (in fact, it now might be possible to start
to use SSE2 to do two rows at once).

Finally, the flags element contains (currently) five bits that
should allow to speed up certain algorithms. The basic idea is
that if flags is zero then we can use the fastest possible algorithm.
The flags contain information about offset, excess bits, it
being a windowed matrix, and the existance of more than one block.