Rewrite of _mzd_addmul_even_weird to use rowstride.
Doesn't seem to speed anything up, but it was a 'test case'
to show how it's done ;). Eliminates the use of 'rows',
reducing the memory access roughly with a factor of two.

Of course, in the light of calling mzd_init, which
still calls malloc for blocks, and rows and fills the
latter with data... this all makes little sense unless
we really get rid of rows (and also cache allocations
of blocks[]).