m4ri.pc.in should include $includedir with /m4ri at the end
m4ri.h expects the user to include ${includedir} but the pkg-config file suggests to include ${includedir}/m4ri which is inconsistent. It is better to include <m4ri/m4ri.h> to avoid conflicts if another project uses the same header filename as one of our headers.

This fixes #60