m4ri.pc.in: move $OPENMP_FLAGS from Cflags to Libs.private
The -fopenmp (or equivalent) flag should not be present in Cflags,
because it causes programs being linked against m4ri to be compiled
with OpenMP support, and that may not be intended. For example, this
causes runtime crashes in SageMath built with meson: while -fopenmp
from m4ri is used at compile time, libgomp is ultimately not linked in
because the build system doesn't think we're using OpenMP.

As a result, $OPENMP_FLAGS flags are more appropriate to include when
linking than compiling. Moreover, the OpenMP library dependency (on
libgomp or equivalent) should be private for the same reason libpng
is: none of its symbols are directly exposed to users of m4ri.